requirements.txt
bleach==6.1.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
PyJWT==2.6.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
tableau-api-lib==0.1.50
tableauserverclient==0.34
typeguard==4.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.0.6
